feat: new function to check for rogue livereload
The 'livereload` subcommand does not kill the running Python process (running on port 3000) with Python 2.7, upon pressing Ctrl-C. In addition, it starts up a new Python process (running on port 5000, with the same PID as the one running on port 3000). 
I created a new function which will check for this rogue Python process if the "FLASK_LIVE_RELOAD" environment variable is set to true and kill it.