feat: password no longer required for sudo in tools container
Previously, if you tried to run the tools-utils.sh debug script from inside a shell in the docker container, it would prompt for sudo password, which the end user will not know b/c we don't set it in the tools container. This change allows the debug script to execute the privileged echo 0 | sudo tee /proc/sys/kernel/yama/ptrace_scope command at any time in the container's lifecycle (not just as entry point/start command) without a disruptive prompt, and does not alter the security or features of the deploy container.