fix: Update the dockerfile apm config so it works correctly
Making a few changes so the APM config works correctly. There were a few issues:

1. the changes to server.env were not valid and likely never worked as expected. It relied on doing variable expansion which isn't supported in server.env
1. it didn't work with the Liberty container in non-root mode which is best practice
1. I'm suspicious of the -Xmx setting the heap to 728Mb. If the docker container has a smaller memory limit this will break.

To make this work I've put the server.env changes in as a docker env var and for the jvm args I've specified them using JVM_ARGS.

I haven't moved the -Xmx setting. I think it is probably better to set the heap usage to use a % of available memory. I think there is a way to do this, but I think this might make sense to put in the base image.